Alteração pra receber um valor float, e condição caso não digite nada pra evitar erros.

+++ app/src/main/java/com/unisc/exercicio_aula_1_1/Tela2.java
@@ -17,8 +17,7 @@ public class Tela2 extends AppCompatActivity {
         setTitle("Resultado da Conversão");
 
         Intent intent = getIntent();
-        int valor = intent.getIntExtra("resposta",0);
-
+        float valor = intent.getFloatExtra("resposta",0);
 
         float conversao = ((valor * 9)/5) + 32;
         Log.d("texto","Resposta: " + conversao);
